feat(5056): move unit tests from dataset to backend-service
