+++ dataprep-backend-service/src/test/java/org/talend/dataprep/dataset/adapter/MockDatasetServer.java
@@ -20,12 +20,12 @@ import java.io.IOException;
 import org.apache.commons.io.IOUtils;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
-import static org.springframework.web.bind.annotation.RequestMethod.GET;
 import static org.talend.dataprep.dataset.adapter.MockDatasetServer.AUTHENTICATION_TOKEN;
 
 @RestController
@@ -34,13 +34,14 @@ public class MockDatasetServer {
 
     public static final String AUTHENTICATION_TOKEN = "authentication-token";
 
-    @RequestMapping(value = "/datasets/{datasetId}", method = GET)
+    @GetMapping("/datasets/{datasetId}")
     public String getById(@RequestParam(required = false) Boolean withUiSpec,
             @RequestParam(required = false) Boolean advanced) throws IOException {
-        return IOUtils.toString(getClass().getResourceAsStream("dataset_payload_example.json"), UTF_8);
+        return IOUtils.toString(getClass().getResourceAsStream(
+                "dataset_payload_example.json"), UTF_8);
     }
 
-    @RequestMapping(value = "/dataset-sample/{datasetId}", method = GET, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    @GetMapping(value = "/dataset-sample/{datasetId}", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
     public String getSampleById(@RequestParam(defaultValue = "0") String offset,
             @RequestParam(defaultValue = "0") String limit) throws IOException {
         return IOUtils.toString(getClass().getResourceAsStream(
