--- dataprep-dataset/src/test/java/org/talend/dataprep/dataset/adapter/TestParent.java
@@ -1,48 +1,0 @@
-/*
- *  ============================================================================
- *
- *  Copyright (C) 2006-2018 Talend Inc. - www.talend.com
- *
- *  This source code is available under agreement available at
- *  https://github.com/Talend/data-prep/blob/master/LICENSE
- *
- *  You should have received a copy of the agreement
- *  along with this program; if not, write to Talend SA
- *  9 rue Pages 92150 Suresnes, France
- *
- *  ============================================================================
- */
-
-package org.talend.dataprep.dataset.adapter;
-
-import org.junit.runner.RunWith;
-import org.springframework.boot.SpringBootConfiguration;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.context.embedded.LocalServerPort;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Import;
-import org.springframework.test.context.junit4.SpringRunner;
-
-import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest(webEnvironment = RANDOM_PORT, classes = TestParent.DatasetClientTestConfiguration.class)
-public abstract class TestParent {
-
-    @LocalServerPort
-    protected int localServerPort;
-
-    @SpringBootConfiguration
-    @EnableAutoConfiguration
-    @Import(org.talend.dataprep.configuration.HttpClient.class)
-    public static class DatasetClientTestConfiguration {
-
-        @Bean
-        MockDatasetServer mockDatasetServer() {
-            return new MockDatasetServer();
-        }
-
-    }
-
-}
